requirements.txt
# Core Machine Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Hugging Face Ecosystem (Models & Data)
diffusers>=0.25.0
datasets>=2.14.0
accelerate>=0.25.0

# Evaluation Metrics
torchmetrics>=1.0.0
scipy>=1.10.0  # Required in the background by torchmetrics for FID math

# Image Processing & Utilities
Pillow>=9.5.0
numpy>=1.24.0,<2.0.0

# Visualization & UI
matplotlib>=3.7.0
tqdm>=4.65.0
gradio==4.44.1
pydantic==2.9.2
fastapi==0.112.2
